+++ stockfish-core/CHANGELOG.md
@@ -12,6 +12,10 @@ adheres to [Semantic Versioning][semver].
 - Fixed errors in CHANGELOG formatting
 - Added some missing `#[inline]` attributes to functions
 
+### Added
+
+- `impl Default for Bitboard`
+
 ## [0.2.0](https://github.com/stouset/stockfish-rs/tree/stockfish-core-v0.2.0/stockfish-core) - 2023-01-16
 
 ### Changed